% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeresolved.R
\name{theme_resolved}
\alias{theme_resolved}
\title{theme_resolved}
\usage{
theme_resolved(base_size = 12, base_family = "Lato")
}
\arguments{
\item{base_size:}{base font size, defaults to 12}

\item{base_family:}{base font family. Defaults to Lato, please use extrafont}
}
\description{
Use this function with theme_set() to set a Resolved style theme for ggplot2
}
\examples{
theme_set(theme_resolved)
}
\keyword{theme,ggplot2}
